+++ example_data/pickle_mat.py
@@ -1,0 +1,6 @@
+import pickle
+
+with open("pickle_example_mat", "rb") as fr:
+    data = pickle.load(fr)
+
+print(data)