+++ src/lib.rs
@@ -87,4 +87,4 @@ pub use special::function::*;
 pub use statistics::ops::*;
 
 #[allow(unused_imports)]
-pub use numerical::ops::*;
+pub use util::pickle::*;