+++ src/numerical/mod.rs
@@ -6,4 +6,3 @@ pub mod bdf;
 pub mod runge_kutta;
 pub mod ode;
 pub mod gauss_legendre;
-pub mod ops;