+++ src/util/mod.rs
@@ -2,3 +2,4 @@ pub mod print;
 pub mod non_macro;
 pub mod useful;
 pub mod api;
+pub mod pickle;